Refactor log watcher helpers (#1483)
