+++ truss/cli/logs/base_watcher.py
@@ -3,29 +3,27 @@ import time
 from abc import ABC, abstractmethod
 from typing import Any, Iterator, List, Optional
 
+from rich import console as rich_console
+
+from truss.cli.logs.utils import ParsedLog, parse_logs
 from truss.remote.baseten.api import BasetenApi
-from truss.shared.types import ParsedLog, SpinnerFactory
 
 # NB(nikhil): This helps account for (1) log processing delays (2) clock skews
 CLOCK_SKEW_BUFFER_MS = 10000
 POLL_INTERVAL_SEC = 2
 
 
-def parse_logs(api_logs: List[Any]) -> List[ParsedLog]:
-    return [ParsedLog.from_raw(api_log) for api_log in api_logs]
-
-
 class LogWatcher(ABC):
     api: BasetenApi
-    spinner_factory: SpinnerFactory
+    console: rich_console.Console
     # NB(nikhil): we add buffer for clock skew, so this helps us detect duplicates.
     # TODO(nikhil): clean up hashes so this doesn't grow indefinitely.
     _log_hashes: set[str] = set()
     _last_poll_time: Optional[int] = None
 
-    def __init__(self, api: BasetenApi, spinner_factory: SpinnerFactory):
+    def __init__(self, api: BasetenApi, console: rich_console.Console):
         self.api = api
-        self.spinner_factory = spinner_factory
+        self.console = console
 
     def _hash_log(self, log: ParsedLog) -> str:
         log_str = f"{log.timestamp}-{log.message}-{log.replica}"
@@ -52,7 +50,7 @@ class LogWatcher(ABC):
 
     def watch(self) -> Iterator[ParsedLog]:
         self.before_polling()
-        with self.spinner_factory("Waiting for logs..."):
+        with self.console.status("Waiting for logs...", spinner="dots"):
             while True:
                 for log in self._poll():
                     yield log
@@ -66,6 +64,8 @@ class LogWatcher(ABC):
             time.sleep(POLL_INTERVAL_SEC)
             self.post_poll()
 
+        self.after_polling()
+
     @abstractmethod
     def fetch_logs(
         self, start_epoch_millis: Optional[int], end_epoch_millis: Optional[int]
@@ -75,6 +75,11 @@ class LogWatcher(ABC):
     @abstractmethod
     def before_polling(self) -> None:
         """Hook to run code before any polling begins."""
+        pass
+
+    @abstractmethod
+    def after_polling(self) -> None:
+        """Hook to run code after all polling ends."""
         pass
 
     @abstractmethod
